+++ R/get_cam_op.R
@@ -14,8 +14,6 @@
 #' in decimal effort values as in [camtrapR::cameraOperation()](
 #' https://jniedballa.github.io/camtrapR/reference/cameraOperation.html).
 #'
-#' @param package Camera trap data package object, as returned by
-#'   `read_camtrap_dp()`.
 #' @param station_col Column name to use for identifying the stations. Default:
 #'   `"locationName"`.
 #' @param camera_col Column name of the column specifying Camera ID. Default:
@@ -28,13 +26,11 @@
 #'   [camtrapR::cameraOperation()](
 #'   https://jniedballa.github.io/camtrapR/reference/cameraOperation.html).
 #'   Default: `FALSE`.
-#' @param datapkg Deprecated. Use `package` instead.
+#' @inheritParams get_species
 #' @param ... filter predicates for filtering on deployments.
 #' @return A matrix. Row names always indicate the station ID. Column names are
 #'   dates.
 #' @family exploration functions
-#' @importFrom dplyr %>% .data
-#' @importFrom rlang !! :=
 #' @export
 #' @examples
 #' library(dplyr)
@@ -67,18 +63,14 @@
 #' 
 #' # Use prefix Station as in camtrapR's camera operation matrix
 #' get_cam_op(mica, use_prefix = TRUE)
-get_cam_op <- function(package = NULL,
+get_cam_op <- function(package,
                        ...,
                        station_col = "locationName",
                        camera_col = NULL,
                        session_col = NULL,
-                       use_prefix = FALSE,
-                       datapkg = lifecycle::deprecated()) {
-  # check camera trap data package
-  check_package(package, datapkg, "get_cam_op")
-  if (is.null(package) & !is.name(datapkg)) {
-    package <- datapkg
-  }
+                       use_prefix = FALSE) {
+  # Check camera trap data package
+  camtrapdp::check_camtrapdp(package)
   
   # Check that station_col is a single string
   assertthat::assert_that(assertthat::is.string(station_col))