+++ R/get_effort.R
@@ -2,8 +2,6 @@
 #'
 #' Gets the effort (deployment duration) per deployment.
 #'
-#' @param package Camera trap data package object, as returned by
-#'   `read_camtrap_dp()`.
 #' @param unit Time unit to use while returning deployment effort (duration).
 #'   One of:
 #'   - `second`
@@ -12,9 +10,8 @@
 #'   - `day`
 #'   - `month`
 #'   - `year`
-#' @param datapkg Deprecated.
-#'   Use `package` instead.
 #' @param ... filter predicates
+#' @inheritParams get_species
 #' @return A tibble data frame with following columns:
 #'   - `deploymentID`: Deployment unique identifier.
 #'   - `effort`: Effort expressed in the unit passed by parameter `unit`.
@@ -23,7 +20,6 @@
 #'   - `effort_duration`: A duration object (duration is a class from lubridate
 #'   package).
 #' @family exploration functions
-#' @importFrom dplyr .data %>%
 #' @export
 #' @examples
 #' # Efforts expressed in hours
@@ -31,37 +27,33 @@
 #'
 #' # Effort expressed as days
 #' get_effort(mica, unit = "day")
-get_effort <- function(package = NULL,
+get_effort <- function(package,
                        ...,
-                       unit = "hour",
-                       datapkg = lifecycle::deprecated()) {
-  # define possible unit values
+                       unit = "hour") {
+  # Define possible unit values
   units <- c("second", "minute", "hour", "day", "month", "year")
 
-  # check unit
+  # Check unit
   check_value(unit, units, "unit", null_allowed = FALSE)
 
-  # check camera trap data package
-  check_package(package, datapkg, "get_effort")
-  if (is.null(package) & !is.name(datapkg)) {
-    package <- datapkg
-  }
+  # Check camera trap data package
+  camtrapdp::check_camtrapdp(package)
   
-  # apply filtering
+  # Apply filtering
   package$data$deployments <- apply_filter_predicate(
     df = package$data$deployments,
     verbose = TRUE, ...
   )
 
-  # get deployments
+  # Get deployments
   deployments <- package$data$deployments
 
-  # calculate effort of deployments
+  # Calculate effort of deployments
   effort_df <-
     deployments %>%
     dplyr::mutate(effort_duration = lubridate::as.duration(.data$end - .data$start)) %>%
     dplyr::select("deploymentID", "effort_duration")
-  # convert effort duration in specified effort time units (arg units)
+  # Convert effort duration in specified effort time units (arg units)
   effort_df$effort <- transform_effort_to_common_units(
     effort = effort_df$effort_duration,
     unit = unit